servers/network_performance_mcp/src/requirements.txt
fastmcp>=1.0.0